Server/requirements.txt
python-dotenv
requests
aiohttp
pandas
aiomysql
aiofiles